Jarvis-ai/requirements.txt
# -----------------------------
# 🧠 AI / LLM
# -----------------------------
openai>=1.0.0

# -----------------------------
# 🎤 Voice (Speech Input)
# -----------------------------
SpeechRecognition>=3.10.0
pyaudio>=0.2.13

# -----------------------------
# 🔊 Voice Output (TTS)
# -----------------------------
pyttsx3>=2.90

# -----------------------------
# 🌐 Web Automation / Scraping
# -----------------------------
requests>=2.31.0
beautifulsoup4>=4.12.2

# -----------------------------
# 💻 System Control
# -----------------------------
psutil>=5.9.8

# -----------------------------
# ⚙️ Environment Config
# -----------------------------
python-dotenv>=1.0.0

# -----------------------------
# 🧪 Testing
# -----------------------------
pytest>=7.4.0

# -----------------------------
# 🧠 Optional (Advanced Upgrades)
# -----------------------------
# Uncomment if you extend features

# numpy>=1.26.0
# pandas>=2.1.0
# fastapi>=0.110.0
# uvicorn>=0.27.0
# pyautogui>=0.9.54
# vosk>=0.3.45
# sounddevice>=0.4.6